refactor: complete profile folder structure
