+++ app/components/complete-profile/candidate/CandidateProfile.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CandidateProfile = () => {
+  return (
+    <div>CandidateProfile</div>
+  )
+}
+
+export default CandidateProfile
