+++ app/components/complete-profile/recruiter/Details.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { useProfileData } from '../../complete-profile/recruiter/CompleteRecruiterProfileContext'
-import Input from "../postjob/location/Input";
-import Button from "../postjob/Button";
+import { useProfileData } from '../../../complete-profile/recruiter/CompleteRecruiterProfileContext';
+import Input from '../../postjob/location/Input';
+import Button from '../../postjob/Button';
 
 const UserType = ()=>
 {
