+++ app/components/complete-profile/recruiter/TermsAndCondition.jsx
